backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AgentChat Backend"
authors = [
    {name = "Agent David",email = "wilian_0104@hotmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "websockets (>=16.0,<17.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "aiosqlite (>=0.22.1,<0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
